#37 package renaming based on the googlegroup discussion
