+++ api/src/main/java/org/eclipse/microprofile/faulttolerance/Timeout.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.eclipse.microprofile.fault.tolerance.inject;
+package org.eclipse.microprofile.faulttolerance;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
