+++ tck/src/main/java/org/eclipse/microprofile/fault/tolerance/tck/retry/clientserver/CircuitBreakerClient.java
@@ -23,7 +23,7 @@ import java.sql.Connection;
 
 import javax.enterprise.context.RequestScoped;
 
-import org.eclipse.microprofile.fault.tolerance.inject.CircuitBreaker;
+import org.eclipse.microprofile.faulttolerance.CircuitBreaker;
 /**
  * A client to demonstrate the maxRetries and max duration configuration
  * @author <a href="mailto:emijiang@uk.ibm.com">Emily Jiang</a>
